Extract src/footer.ts — the status bar, its popovers and the reactor badge
Last of evidence group 1, and the largest: renderFoot + paintFootRl,
closeFootMenus and all four menus it governs (usage, shortcuts,
terminal-engine, reactor), refreshTokens, and setEngine.

Two things worth stating, since neither is obvious from the file list:

The header's reactor badge is in here even though it is not in the footer.
closeFootMenus is the rule that only one floating menu may be open at a
time, and the badge's dropdown is one of them — split them and each module
imports the other's close function. renderAttn/openAttnPop/closeAttnPop and
the $("attnBadge") click therefore moved together with the rest.

setEngine moved from the top of main.ts down beside the popover that is its
only UI. That is the one non-`export` difference from HEAD: a reordering
*within* the new module, not a rewrite — the diff confirms the body is
untouched.

Handlers moved with the module, per the DOM-owning recipe: the enginePop
click, the attnBadge click and the three $("fXSeg") toggles. The
document-level outside-click closes stay in main.ts's global dispatcher and
reach the exported close functions, exactly as caffeinate's already do.

Two seam-rule-2 hooks: setFooterCloseColorPop (the colour popover belongs to
the project rows) and setFooterSetActive (the badge puts a pane on stage).

No tests: render module, per the Phase-1 rule — and its decisions were
already extracted and tested last commit, which is why this one is pure
painting. main.ts 2,234 -> 2,074. Verified mechanical: the relocated block
diffed against `git show HEAD:src/main.ts` differs only by nine `export`
keywords and setEngine's new position. Gates: tsc --noEmit clean, 368
vitest pass. No Rust touched.
