+++ src/main.ts
@@ -13,9 +13,14 @@ import "@xterm/xterm/css/xterm.css";
 import type {
   DiffStat, ExtSession, GitActionResult, Restorable, Runnable, Sess,
 } from "./types";
-import { isAgent, PILL_TEXT, statusKey, type Engine } from "./types";
+import { isAgent, PILL_TEXT, statusKey } from "./types";
 import { $, chord, IS_MAC, MOD, toast } from "./dom";
 import { updateTray } from "./tray";
+import {
+  closeAttnPop, closeEnginePop, closeFootMenus, closeShortPop, closeUsagePop,
+  refreshTokens, renderAttn, renderFoot, setEngine, setFooterCloseColorPop,
+  setFooterSetActive,
+} from "./footer";
 import {
   clearIcon, customIcons, iconFor, pickCustomIcon, probeIcon, resetCustomIcon,
   setIconRenderMini, setIconRenderSidebar,
@@ -33,10 +38,9 @@ import {
   setAppVersion, telem, toggleDbg,
 } from "./debug";
 import {
-  basename, esc, fmtShort, fmtUntil, relTime, setHome, tilde,
+  basename, esc, fmtShort, relTime, setHome, tilde,
 } from "./format";
-import { forecast5h, forecast7d, rl, setRlLogger, type Forecast } from "./rl";
-import { setTokenScanning, tokenScanning, usageRow } from "./usageview";
+import { setRlLogger } from "./rl";
 import { closeCafPop, reconcileCaf, setCafHost } from "./caffeinate";
 import { closeDiff, diffOpen, openDiff, setDiffCloseFootMenus } from "./diffview";
 import {
@@ -51,7 +55,7 @@ import {
   setGitBusy, timelineHtml, verbFor, vitalHtml, wsetHtml,
 } from "./inspectorview";
 import {
-  abbr, applyHook, applyStatusline, permCmd, riskLevel, setOnTurnEnd, setPhase,
+  applyHook, applyStatusline, permCmd, riskLevel, setOnTurnEnd, setPhase,
 } from "./phase";
 import { bumpFrec, frecScore, parsePal, scoreItem, type PalItem } from "./palette";
 import {
@@ -65,11 +69,10 @@ import {
   type SortMode, type WtGroup,
 } from "./state";
 import {
-  allProjects, needsYou, needsYouSessions, nextAfterClose, orderedSessions,
-  reactorLabel, reactorState, urgencyRank,
+  allProjects, needsYou, nextAfterClose, orderedSessions, urgencyRank,
 } from "./grouping";
 import {
-  dormantBusy, extWorking, GCLASS, GLYPH, taskStateText,
+  dormantBusy, extWorking, taskStateText,
 } from "./sidebarview";
 import {
   applyInputs, discoverTasks, execCmd, exitWaiters, lastRunnableById,
@@ -77,9 +80,6 @@ import {
   setTaskLogger, setTaskRepaint, setTaskToast, stopRuleBlocked, stopRules,
   taskPrefs, togglePin, type TaskLaunchOpts,
 } from "./tasks";
-import {
-  setTokenDays, todayKey, tokenDays, tokenScanAt, usage, type DayUsage,
-} from "./usage";
 
 // One-time recovery of localStorage stranded when the app was renamed
 // Muster -> Episko: the bundle id changed (io.respeak.cclauncher ->
@@ -153,14 +153,6 @@ function macShellKeys(id: string): (e: KeyboardEvent) => boolean {
 // termEngine itself lives in ./state (a persisted preference like the sort mode);
 // this is only the validation against what this build actually offers.
 if (!ALL_ENGINES.some((e) => e.id === termEngine)) setTermEngine("embedded");
-function setEngine(id: Engine) {
-  if (id === termEngine) return;
-  setTermEngine(id);
-  localStorage.setItem("cc-term-engine", termEngine);
-  const d = engineDef(id);
-  toast(id === "embedded" ? "New sessions open in the embedded terminal" : `New sessions open in ${d.label} (external)`);
-  renderFoot();
-}
 
 // Persisted theme override (cc-theme). Absent → follow the OS via the CSS
 // `color-scheme` default; an explicit value pins light/dark across restarts.
@@ -193,6 +185,10 @@ setIconRenderMini(renderMini);
 // A finished project reorder reasserts manual sort mode and repaints everything.
 setSidebarSetSort(setSort);
 setSidebarRenderAll(renderAll);
+// The exclusive-popover rule reaches the colour popover, which belongs to the project
+// rows here, and the reactor badge puts a pane on the stage.
+setFooterCloseColorPop(closeColorPop);
+setFooterSetActive(setActive);
 // The settings window changes seven things it does not own; this is the whole of
 // what it can reach back for.
 setSettingsHost({ setTheme, effectiveTheme, setSort, setEngine, bumpFont, applyFontSize, refreshTokens });
@@ -1094,134 +1090,11 @@ function renderInspector(s: Sess | null) {
   if (s.res) html.push(resHtml(s));                              // REFERENCE — cpu/mem, pinned to the bottom
   $("inspector").innerHTML = html.join("");
 }
-function renderFoot() {
-  const total = usage[todayKey()] || 0;
-  $("fSessions").textContent = String(sessions.size);
-  $("fCost").textContent = "$" + total.toFixed(2);
-  paintFootRl("fRl", "fRlReset", forecast5h());
-  paintFootRl("fRl7", "fRl7Reset", forecast7d());
-  $("fEngine").textContent = engineDef(termEngine).label;
-  if ($("usagePop").classList.contains("show")) renderUsagePop();
-}
-// Colour the footer % by its forecast (not its raw level), and show a muted
-// countdown to that window's reset beside it — see the forecast section above.
-function paintFootRl(pctId: string, resetId: string, f: Forecast) {
-  const pctEl = $(pctId), resetEl = $(resetId);
-  pctEl.textContent = f.used != null ? Math.round(f.used) + "%" : "–";
-  pctEl.className = f.used == null ? "" : "s-" + f.status; // neutral until we have a reading
-  resetEl.textContent = f.resetTs != null ? "↻ " + fmtUntil(f.resetTs) : "";
-}
-// The forecast presentation (foreText / verdictChip / usageRow) and the whole
-// Usage analytics panel now live in ./usageview. What stays here is what talks to
-// the DOM or the backend: the popup below, and refreshTokens.
-function renderUsagePop() {
-  const noData = rl.h5 == null && rl.d7 == null;
-  $("usagePop").innerHTML = `<div class="up-h">Claude usage limits</div>
-    ${usageRow("Session", "5-hour window", forecast5h())}
-    ${usageRow("Weekly", "7-day window", forecast7d())}
-    <div class="up-foot"><span>today <b>$${(usage[todayKey()] || 0).toFixed(2)}</b></span><span>${sessions.size} live · account-wide</span></div>
-    ${noData ? `<div class="up-note">Appears once a running session reports a statusLine.</div>` : ""}`;
-}
-function openUsagePop() {
-  const r = $("fUsageSeg").getBoundingClientRect();
-  const pop = $("usagePop");
-  renderUsagePop();
-  closeFootMenus("usagePop");
-  pop.style.left = Math.max(8, Math.min(r.left, window.innerWidth - 260)) + "px";
-  pop.style.bottom = (window.innerHeight - r.top + 6) + "px";
-  pop.style.top = "auto";
-  pop.classList.add("show");
-}
-function closeUsagePop() { $("usagePop").classList.remove("show"); }
+// Two more renderAll() surfaces moved out: the status bar and its popovers (with the
+// header's reactor badge, which shares their exclusive-menu rule) to ./footer, and the
+// menu-bar mirror to ./tray — a native surface, so its repaint is an IPC call rather
+// than an innerHTML assignment, but it hangs off renderAll() like the rest.
 
-
-// Scan the transcripts for token totals, at most once per 10 min (a full read of
-// the recent corpus). Async + cached, so the tab paints instantly from localStorage
-// and re-paints when fresh numbers land. `force` bypasses the throttle.
-async function refreshTokens(force = false) {
-  if (tokenScanning) return;
-  if (!force && tokenDays.length && Date.now() - tokenScanAt < 6e5) return;
-  setTokenScanning(true);
-  if (settingsOpen() && setTab === "usage") renderSettings(); // surface the "scanning…" hint
-  try {
-    setTokenDays(await invoke<DayUsage[]>("token_usage_by_day", { days: 400 }));
-  } catch (e) { dlog("warn", "token scan failed: " + e); }
-  finally { setTokenScanning(false); if (settingsOpen() && setTab === "usage") renderSettings(); }
-}
-
-// Only one floating footer/overlay menu may be open at a time: every open* closes
-// the rest first. (The footer triggers stopPropagation, so the document-level
-// outside-click close never fires for them — this is what keeps them exclusive.)
-function closeFootMenus(keep?: string) {
-  const menus: [string, () => void][] = [
-    ["colorPop", closeColorPop], ["enginePop", closeEnginePop], ["cafPop", closeCafPop],
-    ["usagePop", closeUsagePop], ["attnPop", closeAttnPop], ["shortPop", closeShortPop],
-  ];
-  for (const [id, close] of menus) if (id !== keep) close();
-}
-// Keyboard shortcuts, listed in the footer's ⌘ Shortcuts popover. Keep in sync with
-// the global keydown handler (the sole source of truth for what these actually do).
-const SHORTCUTS: { label: string; chords: string[][] }[] = [
-  { label: "Command palette", chords: [["⌘", "K"]] },
-  { label: "Switch to session 1–9", chords: [["⌘", "1–9"]] },
-  { label: "Open a terminal here", chords: [["⌘", "T"]] },
-  { label: "Toggle sidebar", chords: [["⌘", "B"]] },
-  { label: "Toggle inspector", chords: [["⌘", "I"]] },
-  { label: "Settings", chords: [["⌘", ","]] },
-  { label: "Terminal font size", chords: [["⌘", "+"], ["⌘", "−"], ["⌘", "0"]] },
-];
-function renderShortPop() {
-  const rows = SHORTCUTS.map((s) => {
-    const keys = s.chords
-      .map((c) => `<span class="sc-chord">${c.map((k) => `<kbd>${esc(k)}</kbd>`).join("")}</span>`)
-      .join(`<span class="sc-or">/</span>`);
-    return `<div class="sc-row"><span class="sc-desc">${esc(s.label)}</span><span class="sc-keys">${keys}</span></div>`;
-  }).join("");
-  $("shortPop").innerHTML = `<div class="sc-h">Keyboard shortcuts</div>${rows}`;
-}
-function openShortPop() {
-  const r = $("fShortSeg").getBoundingClientRect();
-  const pop = $("shortPop");
-  renderShortPop();
-  closeFootMenus("shortPop");
-  pop.style.left = Math.max(8, Math.min(r.left, window.innerWidth - 300)) + "px";
-  pop.style.bottom = (window.innerHeight - r.top + 6) + "px";
-  pop.style.top = "auto";
-  pop.classList.add("show");
-}
-function closeShortPop() { $("shortPop").classList.remove("show"); }
-// Header "reactor": one rollup of the fleet's most-urgent state. Clicking it jumps
-// straight to the longest-waiting session in that state (a picker if several).
-function renderAttn() {
-  const list = needsYouSessions();
-  const b = $("attnBadge");
-  if (!list.length) { b.className = "attn-badge"; closeAttnPop(); return; }
-  const dom = reactorState(list[0]);
-  const n = list.filter((s) => reactorState(s) === dom).length;
-  b.className = `attn-badge show react-${dom}${list.length > 1 ? " multi" : ""}`;
-  $("attnBadgeTxt").textContent = reactorLabel(dom, n);
-  if ($("attnPop").classList.contains("show")) { if (list.length > 1) openAttnPop(list); else closeAttnPop(); }
-}
-// Click the reactor → jump to the session; if several need you, a dropdown lists
-// project + title + reason so you can pick which to jump to.
-function badgeLabel(s: Sess) { return s.title || (s.worktree ? `⑃ ${s.branch}` : (s.branch || "session")); }
-function openAttnPop(list: Sess[]) {
-  const r = $("attnBadge").getBoundingClientRect();
-  const pop = $("attnPop");
-  closeFootMenus("attnPop");
-  pop.innerHTML = list.map((s) => {
-    const k = statusKey(s);
-    const reason = s.attention || PILL_TEXT[s.phase];
-    return `<button class="ap-item" data-sel="${s.id}"><span class="ap-dot ${GCLASS[k]}">${GLYPH[k]}</span><span class="ap-main"><span class="ap-proj">${esc(s.project)}</span><span class="ap-ttl">${esc(badgeLabel(s))}</span></span><span class="ap-reason ${GCLASS[k]}">${esc(abbr(reason, 42))}</span></button>`;
-  }).join("");
-  pop.style.right = Math.max(8, window.innerWidth - r.right) + "px";
-  pop.style.left = "auto";
-  pop.style.top = (r.bottom + 6) + "px";
-  pop.classList.add("show");
-}
-function closeAttnPop() { $("attnPop").classList.remove("show"); }
-// The tray mirror moved to ./tray — a native surface, so its repaint is an IPC call
-// rather than an innerHTML assignment, but it hangs off renderAll() like the rest.
 // ▶ Run and ❯ Terminal both act on the active project's directory, so with no
 // session, shell or mirrored external there is nothing for them to act on. Greying
 // them says so up front; a live button whose only response is an error toast reads
@@ -1797,40 +1670,10 @@ document.addEventListener("contextmenu", (e) => {
   openCtxMenu(el.dataset.key, e.clientX, e.clientY);
 });
 
-// ---------- terminal-engine popover (footer "new in …") ----------
-function openEnginePopover() {
-  const seg = $("fEngineSeg");
-  const r = seg.getBoundingClientRect();
-  const pop = $("enginePop");
-  closeFootMenus("enginePop");
-  pop.innerHTML = availEngines.map((id) => {
-    const d = engineDef(id);
-    return `<button class="mp-item ${id === termEngine ? "on" : ""}" data-engine="${id}"><span class="mp-ic">${id === "embedded" ? "▤" : "⧉"}</span><span class="mp-main"><span class="mp-l">${esc(d.label)}</span><span class="mp-s">${esc(d.sub)}</span></span><span class="mp-check">✓</span></button>`;
-  }).join("");
-  pop.style.left = Math.max(8, Math.min(r.left, window.innerWidth - 228)) + "px";
-  pop.style.bottom = (window.innerHeight - r.top + 6) + "px";
-  pop.style.top = "auto";
-  pop.classList.add("show");
-}
-function closeEnginePop() { $("enginePop").classList.remove("show"); }
-$("enginePop").addEventListener("click", (e) => {
-  const b = (e.target as HTMLElement).closest<HTMLElement>("[data-engine]");
-  if (!b) return;
-  setEngine(b.dataset.engine as Engine);
-  closeEnginePop();
-});
-
 // Keep-awake moved to ./caffeinate — it owns its presets, its timer and its
-// popover; reconcileCaf is still called from renderAll below.
-
-// Reactor click → jump straight to the longest-waiting session, or open a picker
-// if several need you.
-$("attnBadge").addEventListener("click", () => {
-  const list = needsYouSessions();
-  if (list.length === 0) return;
-  if (list.length === 1) { setActive(list[0].id); closeAttnPop(); return; }
-  $("attnPop").classList.contains("show") ? closeAttnPop() : openAttnPop(list);
-});
+// popover; reconcileCaf is still called from renderAll below. The terminal-engine
+// popover and the reactor badge's click went to ./footer with the rest of the status
+// bar; the document-level outside-click closes below still reach their close functions.
 
 $("kbar").addEventListener("click", openPalette);
 $("themeBtn").addEventListener("click", toggleTheme);
@@ -1987,9 +1830,6 @@ $("inBody").addEventListener("keydown", (e) => {
 });
 $("setClose").addEventListener("click", closeSettings);
 $("fRepo").addEventListener("click", (e) => { e.preventDefault(); openUrl("https://github.com/respeak-io/episko").catch(() => {}); });
-$("fEngineSeg").addEventListener("click", (e) => { e.stopPropagation(); $("enginePop").classList.contains("show") ? closeEnginePop() : openEnginePopover(); });
-$("fUsageSeg").addEventListener("click", (e) => { e.stopPropagation(); $("usagePop").classList.contains("show") ? closeUsagePop() : openUsagePop(); });
-$("fShortSeg").addEventListener("click", (e) => { e.stopPropagation(); $("shortPop").classList.contains("show") ? closeShortPop() : openShortPop(); });
 $("btnClose").addEventListener("click", () => { if (activeId) closeSession(activeId); });
 
 $("scrim").addEventListener("click", () => { closePalette(); closeWt(); closeDiff(); closeSettings(); closeRunPicker(); closeInputPrompt(); closeTaskManager(); });
